+++ uniffi_bindgen/src/bindings/ruby/mod.rs
@@ -25,7 +25,6 @@ pub fn write_bindings(
     ci: &ComponentInterface,
     out_dir: &Path,
     try_format_code: bool,
-    _is_testing: bool,
 ) -> Result<()> {
     let mut rb_file = PathBuf::from(out_dir);
     rb_file.push(format!("{}.rb", ci.namespace()));